// src/keys/interfaces/authFetch.ts
// Generated by https://quicktype.io

export interface Auth {
    ok:    boolean;
    msg:   string;
    user:  User;
    token: string;
}

// export interface User {
//     _id:       string;
//     name:      string;
//     email:     string;
//     createdAt: string;
//     updatedAt: string;
// }

// Generated by https://quicktype.io

// export interface Auth {
//     ok:   boolean;
//     user: User;
// }

export interface User {
    name:      string;
    email:     string;
    role:      string;
    status:    boolean;
    _id:       string;
    createdAt: string;
    updatedAt: string;
}